output_projects/New-Project_20260329_112055/requirements.txt
streamlit==1.34.0
openai==1.30.3
azure-identity==1.16.0
msal==1.28.0
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
PyPDF2==3.0.1
pandas==2.2.2
plotly==5.22.0
python-dateutil==2.9.0
pytest==8.2.2